% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{borders}
\alias{borders}
\title{borders of the phi delta space}
\usage{
borders(ratio)
}
\arguments{
\item{ratio}{is the ratio of positive and negative of the data.
The default is 1}
}
\value{
a matrix. Each row represents a corner in the following order:
top, right, bottom, left
}
\description{
calculates the corners of the phi delta space
}
\examples{
borders(1.0)
borders(0.5)
borders(2)
}
\author{
rothe
}
