% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{c_matrices}
\alias{c_matrices}
\title{confusion matrices}
\usage{
c_matrices(stats)
}
\arguments{
\item{stats}{c_statistics}
}
\value{
a matrix.  Each column represents a feature.
Each row describes in this order: true negative, FALSE negative, true positive, FALSE negative
}
\description{
calculates the confusion matrices from the c_statistics
}
\examples{
x <- c_statistics(climate_data)
cmat <- c_matrices(x)
}
\author{
rothe
}
