% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{c_statistics}
\alias{c_statistics}
\title{Raw Confusion Statistics}
\usage{
c_statistics(file)
}
\arguments{
\item{file}{raw data from a file, for example the output of \link{read.csv}.
the file must be formarted as follows:
The first column contains tho output of the classifier.  It should only be 1 or 0
The other columns represent the features.
The names of the columns 2.. are considered as the names of the features}
}
\value{
dataframe, first column are the labels, 0 is a negative sample, 1 a positve
the other columns contain the
}
\description{
reformarts the raw file data to c_statistics data
so it can be used for most of the functions in this package.
it can be used directly after loading data from a file like .csv
}
\examples{
data("climate_data")
x <- c_statistics(climate_data)
}
\author{
rothe
}
