% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{calculate entropy}
\usage{
calculate_entropy(spec, sens, ratio = 1)
}
\arguments{
\item{spec}{numeric, is the specificity, the true negative rate}

\item{sens}{numeric, is the sensitivity, the true positive rate}

\item{ratio}{numeric, is the ratio of positive and negative of the data}
}
\value{
entropy of the tuple
}
\description{
calculates the entropy of a specificity and sensitivity tuple
considering the ratio
}
\examples{
calculate_entropy(1,0)
calculate_entropy(0.5,0.6,0.7)
}
\author{
rothe
}
