% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{calculate_ratio}
\alias{calculate_ratio}
\title{calculate ratio}
\usage{
calculate_ratio(stats)
}
\arguments{
\item{stats}{c_statistics}
}
\value{
ratio
}
\description{
calculates the ratio between positive and negative samples
}
\examples{
x <- c_statistics(climate_data)
ratio <- calculate_ratio(x)
}
\author{
rothe
}
