% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{crossings}
\alias{crossings}
\title{Diagram crossings}
\usage{
crossings(ratio, col = "darkblue", ...)
}
\arguments{
\item{ratio}{is the ratio of positive and negative of the data}

\item{col}{the color of the lines.  Default is darkblue}

\item{...}{further graphical parameters, see \link{par}}
}
\description{
adds crossings to the plot depending on the ratio
}
\examples{
x <- c_statistics(climate_data)
ratio <- calculate_ratio(x)
phiDelta_plot_from_data(x, crossing = FALSE)
crossings(ratio, col = "green")
}
\author{
Neumann
}
