% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{dist_to_middle}
\alias{dist_to_middle}
\title{distance to the middle of the space}
\usage{
dist_to_middle(phi, delta, ratio)
}
\arguments{
\item{phi}{numeric value or vector of phi}

\item{delta}{numeric value or vector of delta}

\item{ratio}{is the ratio of positive and negative of the data. The default is 1}
}
\value{
the euclidic distance of the tuple to the middle
}
\description{
calculates the euclidic distance of a phi delta tuple
to the middle of the phi delta space.  This could be used for a rating of the features
}
\examples{
dist_to_middle(1,0,1)
dist_to_middle(0.5,0.3,1)
}
\author{
rothe
}
