% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{dist_to_top}
\alias{dist_to_top}
\title{distance to top or bottom}
\usage{
dist_to_top(phi, delta)
}
\arguments{
\item{phi}{numeric value or vector of phi}

\item{delta}{numeric value or vector of delta}
}
\value{
distance to the top or the bottom corner
}
\description{
calculates the distance of the tuple to the closer
corner of top and bottom of the phi delta space with ratio 1.
This can be used for a ranking of the features
}
\examples{
dist_to_top(1,0)
dist_to_top(0.5,0.3)
}
\author{
rothe
}
