% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{iso_entropy_curve}
\alias{iso_entropy_curve}
\title{isometric entropy}
\usage{
iso_entropy_curve(x, ratio = 1, eps = 0.001, grid_granularity = 0.01)
}
\arguments{
\item{x}{numeric, is the offset for the points}

\item{ratio}{numeric, is the ratio}

\item{eps}{numeric, the epsilon for entropies to be selected}

\item{grid_granularity}{numeric between 0 and 1, defines the granularity of the grid}
}
\description{
draws isometric curves for the entropy by calculating the entropy
for all points in a grid and connecting those within a epsilon enviroment of the value
}
\author{
Neumann
}
