% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{n_matrices}
\alias{n_matrices}
\title{normalized confusion matrices}
\usage{
n_matrices(c_matrices)
}
\arguments{
\item{c_matrices}{confusion matrices}
}
\value{
a matrix.  Each column represents a feature.
Each row describes in this order: true negative rate, FALSE negative rate,
true positive rate, FALSE negative rate
}
\description{
normalizes the confusion matrices
}
\examples{
x <- c_statistics(climate_data)
cmat <- c_matrices(x)
nmat <- n_matrices(cmat)
}
\author{
rothe
}
