% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{phiDelta.convert}
\alias{phiDelta.convert}
\title{Convertion of specificity and sensitivity to phi and delta}
\usage{
phiDelta.convert(spec, sens, ratio = 1)
}
\arguments{
\item{spec}{is the specificity, the true negative rate}

\item{sens}{is the sensitivity, the true positive rate}

\item{ratio}{is the ratio of positive and negative of the data. The default is 1}
}
\value{
List with phi and delta vectors
}
\description{
converts specificity and sensitivity to phi and delta
depending on the ratio
}
\examples{
phiDelta.convert(1,0)
phiDelta.convert(0.5,0.3, ratio = 0.8)
}
\author{
neumann
}
