% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{phiDelta_plot_from_data}
\alias{phiDelta_plot_from_data}
\title{phi delta plot of raw statistic data}
\usage{
phiDelta_plot_from_data(stats, names = NULL, ratio_corrected = TRUE, ...)
}
\arguments{
\item{stats}{matrix of the statistic data of the features and the classifier}

\item{names}{vector with feature names}

\item{ratio_corrected}{logical, if true the plot will concider the ratio
of the positive and negative data samples}

\item{...}{further parameters for the diagram see \link{phiDelta.plot}}
}
\description{
this will create a basic plot directly out of the statistic data (c_statistics)
}
\examples{
x <- c_statistics(climate_data)
phiDelta_plot_from_data(x)
phiDelta_plot_from_data(x, ratio_corrected = FALSE, iso_spec = TRUE, iso_sens = TRUE)
}
\author{
rothe
}
