% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{rank_stats}
\alias{rank_stats}
\title{ranking of the features}
\usage{
rank_stats(stats, ratio_corrected = FALSE, delta_dist = 1)
}
\arguments{
\item{stats}{c_statistics, the data input}

\item{ratio_corrected}{logical, true if ratio shoud be considerd}

\item{delta_dist, }{numeric, the delta value of the anchor for the geometrical ranking
see \link{symmetric_distance}}
}
\description{
this function puts together a number of rankings of the features
}
\author{
rothe
}
