% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_delta.R
\name{symmetric_distance}
\alias{symmetric_distance}
\title{X symmetric distance of a point}
\usage{
symmetric_distance(x, y, anchor)
}
\arguments{
\item{x, y}{numerical, in this case phi and delta but in general the input coordinates}

\item{anchor}{vector (x,y) the anchor for the calculation of the distance}
}
\value{
the smaller distance of (x,y) to eather the positive or negative anchor
}
\description{
calculates the Distance from the positive anchor and the negative anchor
to the point and returns the smaller one.  That means, if y is positive the distance to
the positive anchor will be return, if it is negative,
the negative anchor distance will be calculated
}
\examples{
symmetric_distance(0.5,0.5,c(0,0))
}
